+++ R/EngineGraphQLGitLab.R
@@ -72,11 +72,31 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
      },
 
      # Pull all given files from all repositories of a group.
-     pull_files_from_org = function(org, repos, file_path) {
+     # This is a one query way to get all the necessary info.
+     # However it may fail if query is too complex (too many files in file_paths).
+     # This may be especially the case when trying to pull the data from earlier
+     # pulled files_structure. In such a case GitStats will switch from this function
+     # to iterator over repositories (multiple queries), as it is done for GitHub.
+     get_files_from_org = function(org,
+                                   repos,
+                                   file_paths,
+                                   host_files_structure,
+                                   only_text_files,
+                                   verbose = FALSE,
+                                   progress = verbose) {
        org <- URLdecode(org)
        full_files_list <- list()
        next_page <- TRUE
        end_cursor <- ""
+       if (!is.null(host_files_structure)) {
+         file_paths <- private$get_path_from_files_structure(
+           host_files_structure = host_files_structure,
+           only_text_files = only_text_files,
+           org = org
+         )
+       } else if (is.null(host_files_structure) && only_text_files) {
+         file_paths <- file_paths[!grepl(non_text_files_pattern, file_paths)]
+       }
        while (next_page) {
          files_query <- self$gql_query$files_by_org(
            end_cursor = end_cursor
@@ -86,15 +106,37 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
              gql_query = files_query,
              vars = list(
                "org" = org,
-               "file_paths" = file_path
+               "file_paths" = file_paths
              )
            )
          },
          error = function(e) {
            list()
          })
-         if (length(files_response$data$group) == 0) {
-           cli::cli_alert_danger("Empty")
+         if (private$is_query_error(files_response)) {
+           if (verbose) {
+             purrr::walk(files_response$errors, ~ cli::cli_alert_warning(.))
+           }
+           if (private$is_complexity_error(files_response)) {
+             if (verbose) {
+               cli::cli_alert_info(
+                 cli::col_br_cyan("I will switch to pulling files per repository.")
+               )
+             }
+             full_files_list <- self$get_files_from_org_per_repo(
+               org = org,
+               repos = repos,
+               file_paths = file_paths,
+               host_files_structure = host_files_structure,
+               only_text_files = only_text_files,
+               verbose = verbose,
+               progress = progress
+             )
+             return(full_files_list)
+           }
+         }
+         if (length(files_response$data$group) == 0 && verbose) {
+           cli::cli_alert_danger("Empty response.")
          }
          projects <- files_response$data$group$projects
          files_list <- purrr::map(projects$edges, function(edge) {
@@ -124,7 +166,47 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
        return(full_files_list)
      },
 
-     get_files_structure_from_org = function(org, repos, pattern = NULL, depth = Inf) {
+     # This method is a kind of support to the method above. It is only run when
+     # one query way applied with get_files_from_org() fails due to its complexity.
+     # For more info see docs above.
+     get_files_from_org_per_repo = function(org,
+                                            repos,
+                                            file_paths = NULL,
+                                            host_files_structure = NULL,
+                                            only_text_files = TRUE,
+                                            verbose = FALSE,
+                                            progress = verbose) {
+       if (is.null(repos)) {
+         repo_data <- private$get_repos_data(
+           org = org,
+           repos = repos
+         )
+         repos <- repo_data[["repositories"]]
+       }
+       org_files_list <- purrr::map(repos, function(repo) {
+         if (!is.null(host_files_structure)) {
+           file_paths <- private$get_path_from_files_structure(
+             host_files_structure = host_files_structure,
+             only_text_files = only_text_files,
+             org = org,
+             repo = repo
+           )
+         }
+         files_response <- tryCatch({
+           private$get_file_blobs_response(
+             org = org,
+             repo = repo,
+             file_paths = file_paths
+           )
+         },
+         error = function(e) {
+           list()
+         })
+       }, .progress = progress)
+       return(org_files_list)
+     },
+
+     get_files_structure_from_org = function(org, repos, pattern = NULL, depth = Inf, verbose = FALSE) {
        repo_data <- private$get_repos_data(
          org = org,
          repos = repos
@@ -137,7 +219,7 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
            pattern = pattern,
            depth = depth
          )
-       })
+       }, .progress = verbose)
        names(files_structure) <- repositories
        files_structure <- purrr::discard(files_structure, ~ length(.) == 0)
        return(files_structure)
@@ -162,6 +244,10 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
    ),
 
    private = list(
+
+     is_complexity_error = function(response) {
+       any(purrr::map_lgl(response$errors, ~ grepl("Query has complexity", .$message)))
+     },
 
      # Wrapper over building GraphQL query and response.
      pull_repos_page = function(org = NULL,
@@ -196,23 +282,33 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
        return(result)
      },
 
-     get_file_response = function(org, repo, file_path, files_query) {
-       files_response <- self$gql_response(
-         gql_query = files_query,
+     get_file_blobs_response = function(org, repo, file_paths) {
+       file_blobs_response <- self$gql_response(
+         gql_query = self$gql_query$file_blob_from_repo(),
+         vars = list(
+           "fullPath" = paste0(org, "/", repo),
+           "file_paths" = file_paths
+         )
+       )
+       return(file_blobs_response)
+     },
+
+     get_files_tree_response = function(org, repo, file_path) {
+       files_tree_response <- self$gql_response(
+         gql_query = self$gql_query$files_tree_from_repo(),
          vars = list(
            "fullPath" = paste0(org, "/", repo),
            "file_path" = file_path
          )
        )
-       return(files_response)
+       return(files_tree_response)
      },
 
      get_files_structure_from_repo = function(org, repo, pattern = NULL, depth = Inf) {
-       files_tree_response <- private$get_file_response(
+       files_tree_response <- private$get_files_tree_response(
          org = org,
          repo = repo,
-         file_path = "",
-         files_query = self$gql_query$files_tree_from_repo()
+         file_path = ""
        )
        files_and_dirs_list <- private$get_files_and_dirs(
          files_tree_response = files_tree_response
@@ -228,11 +324,10 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
        while (folders_exist && tier < depth) {
          new_dirs_list <- c()
          for (dir in dirs) {
-           files_tree_response <- private$get_file_response(
+           files_tree_response <- private$get_files_tree_response(
              org = org,
              repo = repo,
-             file_path = dir,
-             files_query = self$gql_query$files_tree_from_repo()
+             file_path = dir
            )
            files_and_dirs_list <- private$get_files_and_dirs(
              files_tree_response = files_tree_response