+++ R/GitHost.R
@@ -130,10 +130,12 @@ GitHost <- R6::R6Class("GitHost",
 
     #' Retrieve content of given text files from all repositories for a host in
     #' a table format.
-    get_files_content = function(file_path, verbose = TRUE) {
+    get_files_content = function(file_path, host_files_structure = NULL, only_text_files = TRUE, verbose = TRUE) {
       files_table <- if (!private$scan_all) {
         private$get_files_content_from_orgs(
           file_path = file_path,
+          host_files_structure = host_files_structure,
+          only_text_files = only_text_files,
           verbose = verbose
         )
       } else {
@@ -781,21 +783,45 @@ GitHost <- R6::R6Class("GitHost",
     },
 
     # Pull files content from organizations
-    get_files_content_from_orgs = function(file_path, verbose) {
+    get_files_content_from_orgs = function(file_path,
+                                           host_files_structure = NULL,
+                                           only_text_files = TRUE,
+                                           verbose = TRUE) {
       graphql_engine <- private$engines$graphql
-      files_table <- purrr::map(private$orgs, function(org) {
+      if (!is.null(host_files_structure)) {
         if (verbose) {
+          cli::cli_alert_info(cli::col_green("I will make use of files structure stored in GitStats."))
+        }
+        result <- private$get_orgs_and_repos_from_files_structure(
+          host_files_structure = host_files_structure
+        )
+        orgs <- result$orgs
+        repos <- result$repos
+      } else {
+        orgs <- private$orgs
+        repos <- private$repos
+      }
+      files_table <- purrr::map(orgs, function(org) {
+        if (verbose) {
+          user_msg <- if (!is.null(host_files_structure)) {
+            "Pulling files from files structure"
+          } else {
+            glue::glue("Pulling files content: [{paste0(file_path, collapse = ', ')}]")
+          }
           show_message(
             host = private$host_name,
             engine = "graphql",
             scope = org,
-            information = glue::glue("Pulling files: [{paste0(file_path, collapse = ', ')}]")
+            information = user_msg
           )
         }
-        graphql_engine$pull_files_from_org(
+        graphql_engine$get_files_from_org(
           org = org,
-          repos = private$repos,
-          file_path = file_path
+          repos = repos,
+          file_paths = file_path,
+          host_files_structure = host_files_structure,
+          only_text_files = only_text_files,
+          verbose = verbose
         ) %>%
           private$prepare_files_table(
             org = org,
@@ -805,6 +831,14 @@ GitHost <- R6::R6Class("GitHost",
         purrr::list_rbind() %>%
         private$add_repo_api_url()
       return(files_table)
+    },
+
+    get_orgs_and_repos_from_files_structure = function(host_files_structure) {
+      result <- list(
+        "orgs" = names(host_files_structure),
+        "repos" = purrr::map(host_files_structure, ~names(.)) %>% unlist() %>% unname()
+      )
+      return(result)
     },
 
     get_files_structure_from_orgs = function(pattern, depth, verbose) {
@@ -827,10 +861,20 @@ GitHost <- R6::R6Class("GitHost",
           org = org,
           repos = private$repos,
           pattern = pattern,
-          depth = depth
+          depth = depth,
+          verbose = verbose
         )
       })
       names(files_structure_list) <- private$orgs
+      files_structure_list <- files_structure_list %>%
+        purrr::discard(~ length(.) == 0)
+      if (length(files_structure_list) == 0 && verbose) {
+        cli::cli_alert_warning(
+          cli::col_yellow(
+            "For {private$host_name} no files structure found."
+          )
+        )
+      }
       return(files_structure_list)
     },
 
@@ -844,8 +888,8 @@ GitHost <- R6::R6Class("GitHost",
           information = glue::glue("Pulling files: [{paste0(file_path, collapse = ', ')}]")
         )
       }
-      files_table <- rest_engine$pull_files(
-        files = file_path,
+      files_table <- rest_engine$get_files(
+        file_paths = file_path,
         verbose = verbose
       ) %>%
         private$prepare_files_table_from_rest() %>%