+++ tests/testthat/test-05-GitHostGitHub.R
@@ -207,8 +207,8 @@ test_that("get_files_structure_from_orgs pulls files structure for repositories 
   )
   expect_snapshot(
     gh_files_structure_from_orgs <- test_host$get_files_structure_from_orgs(
-      pattern = "\\.md|\\.qmd",
-      depth = 1L,
+      pattern = "\\.md|\\.qmd|\\.png",
+      depth = 2L,
       verbose = TRUE
     )
   )
@@ -217,9 +217,53 @@ test_that("get_files_structure_from_orgs pulls files structure for repositories 
     c("r-world-devs", "openpharma")
   )
   purrr::walk(gh_files_structure_from_orgs[[1]], function(repo_files) {
-    expect_true(all(grepl("\\.md|\\.qmd", repo_files)))
+    expect_true(all(grepl("\\.md|\\.qmd|\\.png", repo_files)))
   })
   test_mocker$cache(gh_files_structure_from_orgs)
+})
+
+
+
+test_that("when files_structure is empty, appropriate message is returned", {
+  test_host <- create_github_testhost(
+    repos = c("r-world-devs/GitStats", "openpharma/DataFakeR", "openpharma/VisR"),
+    mode = "private"
+  )
+  expect_snapshot(
+    test_host$get_files_structure_from_orgs(
+      pattern = "\\.png",
+      depth = 1L,
+      verbose = TRUE
+    )
+  )
+})
+
+test_that("get_orgs_and_repos_from_files_structure", {
+  result <- test_host$get_orgs_and_repos_from_files_structure(
+    host_files_structure = test_mocker$use("gh_files_structure_from_orgs")
+  )
+  expect_equal(
+    names(result),
+    c("orgs", "repos")
+  )
+  purrr::walk(result, ~ expect_true(length(.) > 0))
+})
+
+test_that("get_path_from_files_structure gets file path from files structure", {
+  test_graphql_github <- EngineGraphQLGitHub$new(
+    gql_api_url = "https://api.github.com/graphql",
+    token = Sys.getenv("GITHUB_PAT")
+  )
+  test_graphql_github <- environment(test_graphql_github$initialize)$private
+  file_path <- test_graphql_github$get_path_from_files_structure(
+    host_files_structure = test_mocker$use("gh_files_structure_from_orgs"),
+    only_text_files = FALSE,
+    org = "r-world-devs",
+    repo = "GitStats"
+  )
+
+  expect_equal(typeof(file_path), "character")
+  expect_true(length(file_path) > 0)
 })
 
 # public methods